uptime_checker_pubsub/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.11.0
google-auth>=2.17.0
bittensor==9.12.2
attrs (>=25.4.0,<26.0.0)
avro (>=1.12.1,<2.0.0)
bittensor (>=9.12.2,<10.0.0)
babel (>=2.17.0,<3.0.0)
configparser (>=7.2.0,<8.0.0)
fastjsonschema (>=2.21.2,<3.0.0)
frozenlist (>=1.8.0,<2.0.0)
google-api-core (>=2.28.1,<3.0.0)
google-cloud-pubsub (>=2.33.0,<3.0.0)
googleapis-common-protos (>=1.72.0,<2.0.0)
grpc-google-iam-v1 (>=0.14.3,<0.15.0)
httplib2 (>=0.31.0,<0.32.0)
hyperlink (>=21.0.0,<22.0.0)
incremental (>=24.7.2,<25.0.0)
jsonpatch (>=1.33,<2.0)
jsonpointer (>=3.0.0,<4.0.0)
jsonschema (>=4.25.1,<5.0.0)
pyparsing (>=3.2.5,<4.0.0)
pyserial (>=3.5,<4.0)
python-debian (>=1.0.1,<2.0.0)
python-dotenv (>=1.2.1,<2.0.0)
python-magic (>=0.4.27,<0.5.0)
pytz (>=2025.2,<2026.0)
pyyaml (>=6.0.3,<7.0.0)
requests (>=2.32.5,<3.0.0)
urllib3 (>=2.5.0,<3.0.0)
wadllib (>=2.0.0,<3.0.0)
websocket-client (>=1.9.0,<2.0.0)
